requirements.txt
numpy
python-louvain
networkx
matplotlib
seaborn
pandas
tqdm
tensorflow
keras
pillow